Added instructions on where to find properties for classes.
Each property has a description of where it comes from and how to query it. Rules are not documented - should be done later. Types of properties have been changed to represent the data type in the column from which they come. This is not true for columns that can contain strings, but we only need numbers.
Can reduce some 'long's to int, or even byte if we care. Should be discussed.